// src/phaser/models/keys.ts
export enum TextureKeys {
  Background = "background",
  LadyRacer = "lady",
  BlackCar = "black-car",
  BlueCar = "blue-car",
  ChromeCar = "chrome-car",
}

export enum SceneKeys {
  Preloader = "preloader",
  Game = "game",
  GameOver = "game-over",
}

export enum AnimationKeys {
  LadyForward = "lady-forward",
  LadyLeft = "lady-left",
  LadyRight = "lady-right",
}